+++ sdk/template/AzureTemplate/CHANGELOG.md
@@ -1,3 +1,10 @@
 # Release History
-## 1.0.0
+
+# 1.2.0 (Unreleased)
+Azure Template release for Engineering System testing.
+
+## 1.1.0 (2021-05-24)
+Azure Template release for Engineering System testing.
+
+## 1.0.0 (2021-05-21)
 Azure Template release for Engineering System testing.